it lives!!!

+++ Pdfer/Objects/FloatObject.cs
@@ -1,8 +1,8 @@
 namespace Pdfer.Objects;
 
-public class FloatObject(float value) : NumericObject
+public class FloatObject(double value) : NumericObject
 {
-  public float Value => value;
+  public double Value => value;
 
   public override string ToString() => $"{Value}";
 }
